Refactor Fail auth system
